+++ lapce-app/src/editor/movement.rs
@@ -5,7 +5,7 @@ use std::collections::HashSet;
 use lapce_core::{
     buffer::rope_text::RopeText,
     command::MultiSelectionCommand,
-    cursor::{ColPosition, Cursor, CursorMode},
+    cursor::{ColPosition, Cursor, CursorAffinity, CursorMode},
     editor::Editor,
     mode::{Mode, MotionMode, VisualMode},
     movement::{LinePosition, Movement},
@@ -16,7 +16,10 @@ use lapce_core::{
 
 use crate::doc::Document;
 
-use super::view_data::EditorViewData;
+use super::{
+    view_data::EditorViewData,
+    visual_line::{RVLine, VLineInfo},
+};
 
 /// Move a selection region by a given movement.  
 /// Much of the time, this will just be a matter of moving the cursor, but
@@ -24,6 +27,7 @@ use super::view_data::EditorViewData;
 fn move_region(
     view: &EditorViewData,
     region: &SelRegion,
+    affinity: &mut CursorAffinity,
     count: usize,
     modify: bool,
     movement: &Movement,
@@ -57,6 +61,7 @@ fn move_region(
         view,
         region.end,
         region.horiz.as_ref(),
+        affinity,
         count,
         movement,
         mode,
@@ -71,6 +76,7 @@ fn move_region(
 pub fn move_selection(
     view: &EditorViewData,
     selection: &Selection,
+    affinity: &mut CursorAffinity,
     count: usize,
     modify: bool,
     movement: &Movement,
@@ -78,16 +84,19 @@ pub fn move_selection(
 ) -> Selection {
     let mut new_selection = Selection::new();
     for region in selection.regions() {
-        new_selection
-            .add_region(move_region(view, region, count, modify, movement, mode));
+        new_selection.add_region(move_region(
+            view, region, affinity, count, modify, movement, mode,
+        ));
     }
     new_selection
 }
 
+// TODO: It would probably fit the overall logic better if affinity was immutable and it just returned the new affinity!
 pub fn move_offset(
     view: &EditorViewData,
     offset: usize,
     horiz: Option<&ColPosition>,
+    affinity: &mut CursorAffinity,
     count: usize,
     movement: &Movement,
     mode: Mode,
@@ -99,6 +108,7 @@ pub fn move_offset(
             let new_offset = move_left(
                 view.rope_text(),
                 offset,
+                affinity,
                 mode,
                 count,
                 config.editor.atomic_soft_tab_width(),
@@ -108,8 +118,9 @@ pub fn move_offset(
         }
         Movement::Right => {
             let new_offset = move_right(
-                view.rope_text(),
+                view,
                 offset,
+                affinity,
                 mode,
                 count,
                 config.editor.atomic_soft_tab_width(),
@@ -118,44 +129,45 @@ pub fn move_offset(
             (new_offset, None)
         }
         Movement::Up => {
-            let font_size = config.editor.font_size();
             let (new_offset, horiz) =
-                move_up(view, font_size, offset, horiz.cloned(), mode, count);
+                move_up(view, offset, affinity, horiz.cloned(), mode, count);
 
             (new_offset, Some(horiz))
         }
         Movement::Down => {
-            let font_size = config.editor.font_size();
             let (new_offset, horiz) =
-                move_down(view, font_size, offset, horiz.cloned(), mode, count);
+                move_down(view, offset, affinity, horiz.cloned(), mode, count);
 
             (new_offset, Some(horiz))
         }
-        Movement::DocumentStart => (0, Some(ColPosition::Start)),
+        Movement::DocumentStart => {
+            // Put it before any inlay hints at the very start
+            *affinity = CursorAffinity::Backward;
+            (0, Some(ColPosition::Start))
+        }
         Movement::DocumentEnd => {
-            let (new_offset, horiz) = document_end(view.rope_text(), mode);
+            let (new_offset, horiz) = document_end(view.rope_text(), affinity, mode);
 
             (new_offset, Some(horiz))
         }
         Movement::FirstNonBlank => {
-            let (new_offset, horiz) = first_non_blank(view.rope_text(), offset);
+            let (new_offset, horiz) = first_non_blank(view, affinity, offset);
 
             (new_offset, Some(horiz))
         }
         Movement::StartOfLine => {
-            let (new_offset, horiz) = start_of_line(view.rope_text(), offset);
+            let (new_offset, horiz) = start_of_line(view, affinity, offset);
 
             (new_offset, Some(horiz))
         }
         Movement::EndOfLine => {
-            let (new_offset, horiz) = end_of_line(view.rope_text(), offset, mode);
+            let (new_offset, horiz) = end_of_line(view, affinity, offset, mode);
 
             (new_offset, Some(horiz))
         }
         Movement::Line(position) => {
-            let font_size = config.editor.font_size();
             let (new_offset, horiz) =
-                to_line(view, font_size, offset, horiz.cloned(), mode, position);
+                to_line(view, offset, horiz.cloned(), mode, position);
 
             (new_offset, Some(horiz))
         }
@@ -215,6 +227,7 @@ pub fn move_offset(
 fn move_left(
     rope_text: impl RopeText,
     offset: usize,
+    affinity: &mut CursorAffinity,
     mode: Mode,
     count: usize,
     soft_tab_width: Option<usize>,
@@ -232,18 +245,22 @@ fn move_left(
         }
     }
 
+    *affinity = CursorAffinity::Forward;
+
     new_offset
 }
 
 /// Move the offset to the right by `count` amount.
 /// If `soft_tab_width` is `Some` (and greater than 1) then the offset will snap to the soft tab.
 fn move_right(
-    rope_text: impl RopeText,
+    view: &EditorViewData,
     offset: usize,
+    affinity: &mut CursorAffinity,
     mode: Mode,
     count: usize,
     soft_tab_width: Option<usize>,
 ) -> usize {
+    let rope_text = view.rope_text();
     let mut new_offset = rope_text.move_right(offset, mode, count);
 
     if let Some(soft_tab_width) = soft_tab_width {
@@ -257,89 +274,230 @@ fn move_right(
         }
     }
 
+    let (rvline, col) = view.rvline_col_of_offset(offset, *affinity);
+    let info = view.rvline_info(rvline);
+
+    *affinity = if col == info.last_col(&view.text_prov(), false) {
+        CursorAffinity::Backward
+    } else {
+        CursorAffinity::Forward
+    };
+
     new_offset
 }
 
-// These could be abstracted away from `EditorViewData` by having some trait that has a 'get Rope' and 'get point information about text' functions. Would be useful for testing.
-/// Move the offset up by `count` amount.
+fn find_prev_rvline(
+    view: &EditorViewData,
+    start: RVLine,
+    count: usize,
+) -> Option<RVLine> {
+    if count == 0 {
+        return Some(start);
+    }
+
+    // We can't just directly subtract count because of multi-line phantom text.
+    // As just subtracting count wouldn't properly skip over the phantom lines.
+    // So we have to search backwards for the previous line that has real content.
+    let mut info = None;
+    let mut found_count = 0;
+    for prev_info in view.iter_rvlines(true, start).skip(1) {
+        if prev_info.is_empty() {
+            // We skip any phantom text lines in our consideration
+            continue;
+        }
+
+        // Otherwise we found a real line.
+        found_count += 1;
+
+        if found_count == count {
+            // If we've completed all the count instances then we're done
+            info = Some(prev_info);
+            break;
+        }
+        // Otherwise we continue on to find the previous line with content before that.
+    }
+
+    info.map(|info| info.rvline)
+}
+
+/// Move the offset up by `count` amount.  
+/// `count` may be zero, because moving up in a selection just jumps to the start of the selection.
 fn move_up(
     view: &EditorViewData,
-    font_size: usize,
     offset: usize,
+    affinity: &mut CursorAffinity,
     horiz: Option<ColPosition>,
     mode: Mode,
     count: usize,
 ) -> (usize, ColPosition) {
-    let rope_text = view.rope_text();
-
-    let line = rope_text.line_of_offset(offset);
-
-    if line == 0 {
-        let line = rope_text.line_of_offset(offset);
-        let new_offset = rope_text.offset_of_line(line);
+    let rvline = view.rvline_of_offset(offset, *affinity);
+    if rvline.line == 0 && rvline.line_index == 0 {
+        // Zeroth line
         let horiz = horiz.unwrap_or_else(|| {
-            ColPosition::Col(view.line_point_of_offset(offset, font_size).x)
+            ColPosition::Col(view.line_point_of_offset(offset, *affinity).x)
         });
-        return (new_offset, horiz);
+
+        *affinity = CursorAffinity::Backward;
+
+        return (0, horiz);
     }
 
-    let visual_line = view.visual_line(line).saturating_sub(count);
-    let line = view.actual_line(visual_line, false);
+    let Some(rvline) = find_prev_rvline(view, rvline, count) else {
+        // Zeroth line
+        let horiz = horiz.unwrap_or_else(|| {
+            ColPosition::Col(view.line_point_of_offset(offset, *affinity).x)
+        });
+
+        *affinity = CursorAffinity::Backward;
+
+        return (0, horiz);
+    };
 
     let horiz = horiz.unwrap_or_else(|| {
-        ColPosition::Col(view.line_point_of_offset(offset, font_size).x)
+        ColPosition::Col(view.line_point_of_offset(offset, *affinity).x)
     });
-    let col = view.line_horiz_col(line, font_size, &horiz, mode != Mode::Normal);
-    let new_offset = rope_text.offset_of_line_col(line, col);
+    let col = view.rvline_horiz_col(rvline, &horiz, mode != Mode::Normal);
+
+    // TODO: this should maybe be doing `new_offset == info.interval.start`?
+    *affinity = if col == 0 {
+        CursorAffinity::Forward
+    } else {
+        CursorAffinity::Backward
+    };
+
+    let new_offset = view.offset_of_line_col(rvline.line, col);
 
     (new_offset, horiz)
 }
 
-/// Move the offset down by `count` amount.
-fn move_down(
+/// Move down for when the cursor is on the last visual line.
+fn move_down_last_rvline(
     view: &EditorViewData,
-    font_size: usize,
     offset: usize,
+    affinity: &mut CursorAffinity,
     horiz: Option<ColPosition>,
     mode: Mode,
-    count: usize,
 ) -> (usize, ColPosition) {
     let rope_text = view.rope_text();
 
     let last_line = rope_text.last_line();
-    let line = rope_text.line_of_offset(offset);
-    if line == last_line {
-        let new_offset = rope_text.offset_line_end(offset, mode != Mode::Normal);
-        let horiz = horiz.unwrap_or_else(|| {
-            ColPosition::Col(view.line_point_of_offset(offset, font_size).x)
-        });
-        return (new_offset, horiz);
-    }
+    let new_offset = rope_text.line_end_offset(last_line, mode != Mode::Normal);
 
-    let visual_line = view.visual_line(line);
-    let line = view.actual_line(visual_line + count, true);
-    let line = line.min(last_line);
+    // We should appear after any phantom text at the very end of the line.
+    *affinity = CursorAffinity::Forward;
 
     let horiz = horiz.unwrap_or_else(|| {
-        ColPosition::Col(view.line_point_of_offset(offset, font_size).x)
+        ColPosition::Col(view.line_point_of_offset(offset, *affinity).x)
     });
-    let col = view.line_horiz_col(line, font_size, &horiz, mode != Mode::Normal);
-    let new_offset = rope_text.offset_of_line_col(line, col);
 
     (new_offset, horiz)
 }
 
-fn document_end(rope_text: impl RopeText, mode: Mode) -> (usize, ColPosition) {
+fn find_next_rvline_info(
+    view: &EditorViewData,
+    offset: usize,
+    start: RVLine,
+    count: usize,
+) -> Option<VLineInfo<()>> {
+    // We can't just directly add count because of multi-line phantom text.
+    // These lines are 'not there' and also don't have any position that can be moved into
+    // (unlike phantom text that is mixed with real text)
+    // So we have to search forward for the next line that has real content.
+    // The typical iteration count for this is 1, and even after that it is usually only a handful.
+    let mut found_count = 0;
+    for next_info in view.iter_rvlines(false, start) {
+        if count == 0 {
+            return Some(next_info);
+        }
+
+        if next_info.is_empty() {
+            // We skip any phantom text lines in our consideration
+            // TODO: Would this skip over an empty line?
+            continue;
+        }
+
+        if next_info.interval.start <= offset {
+            // If we're on or before our current visual line then we skip it
+            continue;
+        }
+
+        // Otherwise we found a real line.
+        found_count += 1;
+
+        if found_count == count {
+            // If we've completed all the count instances then we're done
+            return Some(next_info);
+        }
+        // Otherwise we continue on to find the next line with content after that.
+    }
+
+    None
+}
+
+/// Move the offset down by `count` amount.  
+/// `count` may be zero, because moving down in a selection just jumps to the end of the selection.
+fn move_down(
+    view: &EditorViewData,
+    offset: usize,
+    affinity: &mut CursorAffinity,
+    horiz: Option<ColPosition>,
+    mode: Mode,
+    count: usize,
+) -> (usize, ColPosition) {
+    let rvline = view.rvline_of_offset(offset, *affinity);
+
+    let Some(info) = find_next_rvline_info(view, offset, rvline, count) else {
+        // There was no next entry, this typically means that we would go past the end if we went
+        // further
+        return move_down_last_rvline(view, offset, affinity, horiz, mode);
+    };
+
+    // TODO(minor): is this the right affinity?
+    let horiz = horiz.unwrap_or_else(|| {
+        ColPosition::Col(view.line_point_of_offset(offset, *affinity).x)
+    });
+
+    let col = view.rvline_horiz_col(info.rvline, &horiz, mode != Mode::Normal);
+
+    let new_offset = view.offset_of_line_col(info.rvline.line, col);
+
+    *affinity = if new_offset == info.interval.start {
+        // The column was zero so we shift it to be at the line itself.
+        // This lets us move down to an empty - for example - next line and appear at the
+        // start of that line without coinciding with the offset at the end of the previous line.
+        CursorAffinity::Forward
+    } else {
+        CursorAffinity::Backward
+    };
+
+    (new_offset, horiz)
+}
+
+fn document_end(
+    rope_text: impl RopeText,
+    affinity: &mut CursorAffinity,
+    mode: Mode,
+) -> (usize, ColPosition) {
     let last_offset =
         rope_text.offset_line_end(rope_text.len(), mode != Mode::Normal);
+
+    // Put it past any inlay hints directly at the end
+    *affinity = CursorAffinity::Forward;
 
     (last_offset, ColPosition::End)
 }
 
-fn first_non_blank(rope_text: impl RopeText, offset: usize) -> (usize, ColPosition) {
-    let line = rope_text.line_of_offset(offset);
-    let non_blank_offset = rope_text.first_non_blank_character_on_line(line);
-    let start_line_offset = rope_text.offset_of_line(line);
+fn first_non_blank(
+    view: &EditorViewData,
+    affinity: &mut CursorAffinity,
+    offset: usize,
+) -> (usize, ColPosition) {
+    let info = view.rvline_info_of_offset(offset, *affinity);
+    let non_blank_offset = info.first_non_blank_character(&view.text_prov());
+    let start_line_offset = info.interval.start;
+    // TODO: is this always the correct affinity? It might be desirable for the very first character on a wrapped line?
+    *affinity = CursorAffinity::Forward;
+
     if offset > non_blank_offset {
         // Jump to the first non-whitespace character if we're strictly after it
         (non_blank_offset, ColPosition::FirstNonBlank)
@@ -354,26 +512,41 @@ fn first_non_blank(rope_text: impl RopeText, offset: usize) -> (usize, ColPositi
     }
 }
 
-fn start_of_line(rope_text: impl RopeText, offset: usize) -> (usize, ColPosition) {
-    let line = rope_text.line_of_offset(offset);
-    let new_offset = rope_text.offset_of_line(line);
+fn start_of_line(
+    view: &EditorViewData,
+    affinity: &mut CursorAffinity,
+    offset: usize,
+) -> (usize, ColPosition) {
+    let rvline = view.rvline_of_offset(offset, *affinity);
+    let new_offset = view.offset_of_rvline(rvline);
+    // TODO(minor): if the line has zero characters, it should probably be forward affinity but
+    // other cases might be better as backwards?
+    *affinity = CursorAffinity::Forward;
 
     (new_offset, ColPosition::Start)
 }
 
 fn end_of_line(
-    rope_text: impl RopeText,
+    view: &EditorViewData,
+    affinity: &mut CursorAffinity,
     offset: usize,
     mode: Mode,
 ) -> (usize, ColPosition) {
-    let new_offset = rope_text.offset_line_end(offset, mode != Mode::Normal);
+    let info = view.rvline_info_of_offset(offset, *affinity);
+    let new_col = info.last_col(&view.text_prov(), mode != Mode::Normal);
+    *affinity = if new_col == 0 {
+        CursorAffinity::Forward
+    } else {
+        CursorAffinity::Backward
+    };
+
+    let new_offset = view.offset_of_line_col(info.rvline.line, new_col);
 
     (new_offset, ColPosition::End)
 }
 
 fn to_line(
     view: &EditorViewData,
-    font_size: usize,
     offset: usize,
     horiz: Option<ColPosition>,
     mode: Mode,
@@ -381,23 +554,24 @@ fn to_line(
 ) -> (usize, ColPosition) {
     let rope_text = view.rope_text();
 
+    // TODO(minor): Should this use rvline?
     let line = match position {
         LinePosition::Line(line) => (line - 1).min(rope_text.last_line()),
         LinePosition::First => 0,
         LinePosition::Last => rope_text.last_line(),
     };
+    // TODO(minor): is this the best affinity?
     let horiz = horiz.unwrap_or_else(|| {
-        ColPosition::Col(view.line_point_of_offset(offset, font_size).x)
+        ColPosition::Col(
+            view.line_point_of_offset(offset, CursorAffinity::Backward)
+                .x,
+        )
     });
-    let col = view.line_horiz_col(line, font_size, &horiz, mode != Mode::Normal);
+    let col = view.line_horiz_col(line, &horiz, mode != Mode::Normal);
     let new_offset = rope_text.offset_of_line_col(line, col);
 
     (new_offset, horiz)
 }
-
-// TODO: passing a view with the document is kinda ehhh, because this would be used when you call
-// .update on the `RwSignal<Document>`, which presumably delinks it from the RwSignal the view has.
-// A solution might be to change the caller to only pass in the editor view and have this do the update, but only when it is needed.
 
 /// Move the current cursor.  
 /// This will signal-update the document for some motion modes.
@@ -420,6 +594,7 @@ pub fn move_cursor(
                 view,
                 offset,
                 cursor.horiz.as_ref(),
+                &mut cursor.affinity,
                 count,
                 movement,
                 Mode::Normal,
@@ -429,6 +604,7 @@ pub fn move_cursor(
                     view,
                     new_offset,
                     None,
+                    &mut cursor.affinity,
                     1,
                     &Movement::Right,
                     Mode::Insert,
@@ -473,6 +649,7 @@ pub fn move_cursor(
                 view,
                 end,
                 cursor.horiz.as_ref(),
+                &mut cursor.affinity,
                 count,
                 movement,
                 Mode::Visual(VisualMode::Normal),
@@ -488,6 +665,7 @@ pub fn move_cursor(
             let selection = move_selection(
                 view,
                 selection,
+                &mut cursor.affinity,
                 count,
                 modify,
                 movement,
@@ -522,6 +700,7 @@ pub fn do_multi_selection(
                     view,
                     offset,
                     cursor.horiz.as_ref(),
+                    &mut cursor.affinity,
                     1,
                     &Movement::Up,
                     Mode::Insert,
@@ -540,6 +719,7 @@ pub fn do_multi_selection(
                     view,
                     offset,
                     cursor.horiz.as_ref(),
+                    &mut cursor.affinity,
                     1,
                     &Movement::Down,
                     Mode::Insert,
@@ -747,4 +927,5 @@ pub fn do_motion_mode(
     }
 }
 
-// TODO: Write tests for the various functions.
+// TODO: Write tests for the various functions. We'll need a more easily swappable API than
+// `EditorViewData` for that.