+++ lapce-app/src/window_tab.rs
@@ -21,8 +21,8 @@ use floem::{
 use indexmap::IndexMap;
 use itertools::Itertools;
 use lapce_core::{
-    command::FocusCommand, directory::Directory, meta, mode::Mode,
-    register::Register,
+    command::FocusCommand, cursor::CursorAffinity, directory::Directory, meta,
+    mode::Mode, register::Register,
 };
 use lapce_rpc::{
     core::CoreNotification,
@@ -1797,8 +1797,11 @@ impl WindowTabData {
         let (window_origin, viewport, view) =
             (editor.window_origin, editor.viewport, editor.view.clone());
 
-        let (point_above, point_below) =
-            view.points_of_offset(self.common.hover.offset.get_untracked());
+        // TODO(minor): affinity should be gotten from where the hover was started at.
+        let (point_above, point_below) = view.points_of_offset(
+            self.common.hover.offset.get_untracked(),
+            CursorAffinity::Forward,
+        );
 
         let window_origin =
             window_origin.get() - self.common.window_origin.get().to_vec2();
@@ -1840,7 +1843,10 @@ impl WindowTabData {
         let (window_origin, viewport, view) =
             (editor.window_origin, editor.viewport, editor.view.clone());
 
-        let (point_above, point_below) = view.points_of_offset(completion.offset);
+        // TODO(minor): What affinity should we use for this? Probably just use the cursor's
+        // original affinity..
+        let (point_above, point_below) =
+            view.points_of_offset(completion.offset, CursorAffinity::Forward);
 
         let window_origin =
             window_origin.get() - self.common.window_origin.get().to_vec2();
@@ -1890,7 +1896,9 @@ impl WindowTabData {
         let (window_origin, viewport, view) =
             (editor.window_origin, editor.viewport, editor.view.clone());
 
-        let (_point_above, point_below) = view.points_of_offset(code_action.offset);
+        // TODO(minor): What affinity should we use for this?
+        let (_point_above, point_below) =
+            view.points_of_offset(code_action.offset, CursorAffinity::Forward);
 
         let window_origin =
             window_origin.get() - self.common.window_origin.get().to_vec2();
@@ -1940,8 +1948,11 @@ impl WindowTabData {
         let (window_origin, viewport, view) =
             (editor.window_origin, editor.viewport, editor.view.clone());
 
-        let (_point_above, point_below) =
-            view.points_of_offset(self.rename.start.get_untracked());
+        // TODO(minor): What affinity should we use for this?
+        let (_point_above, point_below) = view.points_of_offset(
+            self.rename.start.get_untracked(),
+            CursorAffinity::Forward,
+        );
 
         let window_origin =
             window_origin.get() - self.common.window_origin.get().to_vec2();